requirements.txt
fastapi[all]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
notion-client
pydantic
supabase
cryptography
python-dotenv
